cms-api: Expose DAM file size as Float to fix overflow for files >2 GB
The size field of DamFile and FileUpload is stored as a bigint in the
database but was exposed in GraphQL as Int, which only covers 32-bit
signed integers (max ~2.1 GB). Files larger than that (e.g. videos)
caused the GraphQL response to fail with "Int cannot represent non
32-bit signed integer value", breaking the DAM file list and the
"Select files from DAM" dialog. Expose it as Float instead, which
safely represents the full range of realistic file sizes.

+++ packages/api/cms-api/src/dam/files/entities/file.entity.ts
@@ -14,7 +14,7 @@ import {
     Property,
 } from "@mikro-orm/postgresql";
 import { Type } from "@nestjs/common";
-import { Field, ID, Int, ObjectType } from "@nestjs/graphql";
+import { Field, Float, ID, ObjectType } from "@nestjs/graphql";
 import { v4 as uuid } from "uuid";
 
 import { EntityInfo } from "../../../entity-info/entity-info.decorator";
@@ -73,7 +73,7 @@ export function createFileEntity({ Scope, Folder }: { Scope?: Type<DamScopeInter
         @Property({ columnType: "text" })
         name: string;
 
-        @Field(() => Int)
+        @Field(() => Float)
         @Property({ type: new BigIntType("number") })
         size: number;
